day46/day46_windowed_analytics/requirements.txt
redis==5.0.4
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
pytest==8.2.2
pytest-asyncio==0.23.7
websockets==12.0
aioredis==2.0.1
pandas==2.2.2
plotly==5.20.0
jinja2==3.1.4
python-multipart==0.0.9
asyncio==3.4.3
structlog==24.1.0
python-dateutil==2.9.0
pytz==2024.1
numpy==1.26.4
aiofiles==23.2.1
httpx==0.27.0
